requirements.txt
python-dotenv
pandas
langchain
langchain-google-genai
tavily-python
